Fix the issue with displaying separate hidden assets
* adding new assets option should be displayed only once
* add a dividing line
